Add latest reveal.js library in separate directory
Each presentation will need access the reveal.js framework library, both when creating the final presentation as well as when presenting it. This could also be accomplished by a remote reference, but then a local version of the presentation requires an internet connection **and** certain features (presenter mode) would not work.

Thus, this template opts for a self-contained approach and include the complete reveal.js library.