Add barebones example presentation and documented workflow
This commit provides a very simple example and documents the artifacts and workflow in the README. All key elements (including inline, executable R code) are documented in the sample presentation. It is thus a workable first version, but does not contain any major customization beyond the option of choosing a different theme.

The Pandoc-generated presentation `index.html` can be served up locally or via GitHub pages.