Add basic support for a custom Pandoc reveal.js template
The reveal.js framework provides some recent features (for example the `controlsLayout` option) that are not covered by the current default Pandoc template. The only way to enable this is to modify the template.
This commit only adds a copy of the default Pandoc reveal.js template as well as instructions on how to use it with the `pandoc` command.