Add support for custom CSS stylesheets and some examples
This commit provides a custom CSS stylesheet and instructions (in the README) how to use it. This initial version of the custom CSS also has some style changes that affect presentation-wide style as well as some typical customizations that would be applied on individual slides (slides without headers, slides with blockquotes).