Add MathJax support for mathematical formulas
Includes an example slide of a mathematical formula and provides changes to support offline rendering with MathJax.
A local copy of MathJax is added to the repo, the custom Pandoc template is modified to reference this local copy and use an SVG MathJax output processor. The README is updated with a documentation section.
License notices with reference to MathJax are updated in the README and the sample presentation.